Move position conversion stuff to util module
So that the Scala 3 compiler module can use it too
