+++ amm/interp/src/main/scala/ammonite/interp/script/Diagnostic.scala
@@ -1,5 +1,7 @@
 package ammonite.interp.script
 
+import ammonite.util.Position
+
 final case class Diagnostic(
   severity: String,
   start: Position,
