+++ amm/util/src/main/scala/ammonite/util/Position.scala
@@ -1,3 +1,3 @@
-package ammonite.interp.script
+package ammonite.util
 
 final case class Position(line: Int, char: Int)
